% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Auxiliary.R
\name{MessageToHtml}
\alias{MessageToHtml}
\title{MessageToHtml}
\usage{
MessageToHtml(message)
}
\arguments{
\item{message}{\code{character vector} of length 1 with optional name}
}
\description{
Turn message into HTML
}
