% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateTable.Counter.R
\name{CreateTable.Counter}
\alias{CreateTable.Counter}
\title{CreateTable.Counter}
\usage{
CreateTable.Counter(CounterData, ShowAllStages)
}
\arguments{
\item{CounterData}{\code{list} - Contains...}

\item{ShowAllStages}{\code{logical flag} - Whether to show columns regarding changes in different processing stages}
}
\value{
Table object of \code{reactable} class
}
\description{
Compile table check data from \code{dsFredaClient::ds.CheckTable()} or \code{dsFredaClient::ds.CheckDataSet()} into a table suited for display
}
\author{
Bastian Reiter
}
